controller: Introduce Disk Controller and update other controllers
1. Disk controller is responsible for verifying the data info, updating the disk
status and the related replicas.
  1. Clean up most of the status fields when a disk becomes 'disconnected'.
  2. When the disk UUID which is already tracked by an existing disconnected disk is
found in a new disk data path, it means the actual disk is migrated, and the actual
disk will be tracked by the new disk object. In this case, the status.DiskUUID of
the old disk object will be cleaned up. And the related replicas should use new
DiskID with new data path.
  3. When the running replicas are still trying to use the data path of a disconnected
disk, disk controller should set Spec.FailedAt for these replicas immediately.
2. Remove the disk related logic from node controller.

Longhorn #1269

Signed-off-by: Shuo Wu <shuo@rancher.com>