Fix: volume controller constantly update the ScheduledCondition
When a volume is created with number of replicas more than available nodes
and AllowVolumeCreationWithDegradedAvailability is enabled, Volume controller
constantly updates the volume's ScheduledCondition. This prevents other components
to update volume's spec since there will be a lot of conflict. As a result,
volume sometime takes a longhorn time to attach.

This PR make sure volume controller only update ScheduledCondition if it needs
to change.

Longhorn #2073

Signed-off-by: Phan Le <phan.le@rancher.com>